CIS fix for unsigned/signed mismatch in sparse dynamic point octree #jira UE-131228 #rb none #lockdown mitchell.wilson
[CL 17774237 by Jimmy Andrews in 4.27-Chaos branch]

+++ Engine/Plugins/Experimental/GeometryProcessing/Source/GeometricObjects/Public/Spatial/SparseDynamicPointOctree3.h
@@ -329,7 +329,7 @@ protected:
 
 	uint32 GetCellForPoint(int32 PointID) const
 	{
-		if (PointID >= 0 && PointID < PointIDToCellMap.Num())
+		if (PointID >= 0 && size_t(PointID) < PointIDToCellMap.Num())
 		{
 			return PointIDToCellMap[PointID];
 		}
